% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid-list.R
\name{is_hybrid_list}
\alias{is_hybrid_list}
\title{Is \code{x} a hybrid list?}
\usage{
is_hybrid_list(x)
}
\description{
A hybrid list is a wrapper around the list of chunks that make up a single
chunked column.
}
