% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid-replace.R
\name{is_replaceable_mean}
\alias{is_replaceable_mean}
\title{Is this a simple mean call?}
\usage{
is_replaceable_mean(fn, args)
}
\description{
Yes
mean(hybrid_list)
}
\details{
Yes
mean(hybrid_list, na.rm = TRUE)

No
mean(S3)

No
mean(hybrid_list_of_s3)

No
mean(hybrid_list, hybrid_list)
}
